impr: split direction helper service into three classes
